Kludge: Set Graphic.height/.width with no RA.
The graphic should extend to the end of the screen (bottom right corner),
(unless transparent):. However, the Graphic struct presumes that the bitmap
size is a multiple of the pixh/pixw aspect ratio. This is silly and could be
worked around, but the proper solution is probably to just not use graphic.c
at all.